Change res.attachment to setHeader with Content-Disposition, and filename without quotes. This appears to work better when downloading the files.
(With res.attachment they came back with quotes around the filename,
and for some reason this cause all tested browsers to thing the quotes
are part of the filename, substituting them with underscores, also
messing up any filetype in the filename. This appears to go against
the spec, quotes are optional, and in some circumstances required, so
why?)

Updating some dependencies.